コード整理。デバッグ用の時間計測を分離
