+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContextExtensions.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using UniGLTF.AltTask;
+using UnityEngine;
 
 namespace UniGLTF
 {
@@ -32,9 +33,10 @@ namespace UniGLTF
         /// </summary>
         public static void Load(this ImporterContext self)
         {
+            var meassureTime = new ImporterContextSpeedLog();
             using (var queue = TaskQueue.Create())
             {
-                var task = self.LoadAsync();
+                var task = self.LoadAsync(meassureTime.MeasureTime);
 
                 // 中断された await を消化する
                 while (!task.IsCompleted)
@@ -43,6 +45,10 @@ namespace UniGLTF
                     queue.ExecuteOneCallback();
                 }
             }
+
+#if VRM_DEVELOP
+            Debug.Log(meassureTime.GetSpeedLog());
+#endif
         }
     }
 }
