+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContextSpeedLog.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e99bb16028e30ce4b972caecca5343b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
